added some JS libraries
we now use json2html lib to generate our HTML from the JSON, this
simplifies a lot as we just need to specify a template - which supports
imbedded JS and `events'.

We also are able to properly save a JSON file, using the FileSaver lib
(assumes that our browser supports BLOB... anyone using IE?).

Minor changes also to the JS.